+++ resources/js/components/AttributesPanel.vue
@@ -98,11 +98,11 @@ export default {
     },
 
     created() {
-        this.bard.$on('bts-update', () => this.$emit('close'));
+        this.bard.$on('bts-update', this.close);
     },
 
     beforeDestroy() {
-        this.bard.$off('bts-update');
+        this.bard.$off('bts-update', this.close);
     },
 
     methods: {
@@ -117,6 +117,10 @@ export default {
                 items: this.items,
             });
             this.$emit('applied');
+        },
+
+        close() {
+            this.$emit('close');
         },
 
         toggleTrueValue(field) {