+++ CHANGELOG.md
@@ -2,7 +2,14 @@
 
 ## 2.2.0 ()
 
-- [new] Statamic 4 support and updated button icons
+- [new] Statamic 4 support (plus updated button icons)
+- []
+- []
+- []
+- []
+- []
+- []
+- []
 
 ## 2.1.1 (2023-02-16)