+++ CHANGELOG.md
@@ -3,7 +3,14 @@
 ## 2.2.0 ()
 
 - [new] Statamic 4 support and updated button icons
-- [new] Select inputs in the attributes panel
+- [new] Toggle attribute values can now be configured
+- [new] Select type for attributes
+- []
+- []
+- []
+- []
+- []
+- Non-rendered toggle attribute values are now `true`/`false` 
 
 ## 2.1.1 (2023-02-16)