+++ CHANGELOG.md
@@ -4,13 +4,13 @@
 
 - [new] Statamic 4 support and updated button icons
 - [new] Toggle attribute values can now be configured
+- [new] Select type for attributes
 - []
 - []
 - []
 - []
 - []
-- []
-- Non-rendered toggle attributes are now `true`/`false` 
+- Non-rendered toggle attribute values are now `true`/`false` 
 
 ## 2.1.1 (2023-02-16)