+++ CHANGELOG.md
@@ -2,8 +2,16 @@
 
 ## 2.2.0 ()
 
-- [new] Statamic 4 support and updated button icons
-- [new] Option to show current style name in the style selection button 
+- [new] Statamic 4 support (with updated button icons)
+- [new] Option to show style name in the style selection button
+- [new] Toggle attribute values can now be configured
+- [new] Select type for attributes
+- []
+- []
+- []
+- []
+- []
+- Non-rendered toggle attribute values are now `true`/`false` 
 
 ## 2.1.1 (2023-02-16)